DL-13488, rawePageHistory optional in Session, moved model classes out of repositories package
