+++ app/models/GuidanceSession.scala
@@ -19,7 +19,6 @@ package models
 import core.models.ocelot.stanzas.Stanza
 import core.models.ocelot.{FlowStage, SecuredProcess, Process, Label, RunMode, Published}
 import play.api.libs.json.{Json, OFormat}
-import repositories.{PageHistory, Session}
 
 case class GuidanceSession(process: Process,
                            answers: Map[String, String],
