+++ app/models/PageHistory.scala
@@ -16,7 +16,7 @@
 
 // $COVERAGE-OFF$
 
-package repositories
+package models
 
 import play.api.libs.functional.syntax._
 import play.api.libs.json.Reads._
